third_party/WebKit/LayoutTests/fast/events/resize-events-count-expected.txt

Test how many resize events are emitted during window resize.
PASS resizeEventCount is 1
